Initialize sessionDirectory and targetFile when the manager is created and not in initConfig
